YEL-17 [refactor] Entity와 Repository를 동일한 패키지에서 관리하도록 변경
