+++ src/main/java/com/yello/server/domain/cooldown/entity/CooldownRepository.java
@@ -1,7 +1,7 @@
-package com.yello.server.domain.cooldown.repository;
+package com.yello.server.domain.cooldown.entity;
 
-import com.yello.server.domain.cooldown.entity.Cooldown;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CooldownRepository extends JpaRepository<Cooldown, Long> {
+
 }
