+++ src/main/java/com/yello/server/domain/friend/entity/FriendRepository.java
@@ -1,7 +1,7 @@
-package com.yello.server.domain.friend.repository;
+package com.yello.server.domain.friend.entity;
 
-import com.yello.server.domain.friend.entity.Friend;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FriendRepository extends JpaRepository<Friend, Long> {
+
 }
