+++ src/main/java/com/yello/server/domain/group/entity/SchoolRepository.java
@@ -1,7 +1,7 @@
-package com.yello.server.domain.group.repository;
+package com.yello.server.domain.group.entity;
 
-import com.yello.server.domain.group.entity.School;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SchoolRepository extends JpaRepository<School, Long> {
+
 }
