+++ src/main/java/com/yello/server/domain/keyword/entity/KeywordRepository.java
@@ -1,7 +1,7 @@
-package com.yello.server.domain.keyword.repository;
+package com.yello.server.domain.keyword.entity;
 
-import com.yello.server.domain.keyword.entity.Keyword;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface KeywordRepository extends JpaRepository<Keyword, Long> {
+
 }
