+++ src/main/java/com/yello/server/domain/question/entity/QuestionRepository.java
@@ -1,7 +1,7 @@
-package com.yello.server.domain.question.repository;
+package com.yello.server.domain.question.entity;
 
-import com.yello.server.domain.question.entity.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface QuestionRepository extends JpaRepository<Question, Long> {
+
 }
