+++ src/main/java/com/yello/server/domain/user/entity/UserRepository.java
@@ -1,7 +1,7 @@
-package com.yello.server.domain.user.repository;
+package com.yello.server.domain.user.entity;
 
-import com.yello.server.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
+
 }
