+++ src/main/java/com/yello/server/domain/vote/entity/VoteRepository.java
@@ -1,7 +1,7 @@
-package com.yello.server.domain.vote.repository;
+package com.yello.server.domain.vote.entity;
 
-import com.yello.server.domain.vote.entity.Vote;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface VoteRepository extends JpaRepository<Vote, Long> {
+
 }
